Fix tests after df80a7a511aafc78c38e3428556f70a4f3bb2d31 broke them.
Instead of assuming that is_link_to_canvas is false (which is not the case
for many of the tests) simply because it's not a POST-request (as Rack
1.1.0 does not allow @request.request_parameters for non-POST requests),
fix the code properly by dropping the request_parameters call.

The params argument of link_to_canvas? was indeed not used at all, so
simply drop it out and pass only the options hash to the function.